requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
SQLAlchemy==2.0.30
alembic==1.13.1
pandas==2.2.2
onnxruntime==1.17.1
numpy==1.26.4
jinja2==3.1.4
python-multipart==0.0.9
structlog==24.1.0
prometheus-client==0.20.0
orjson
tokenizers==0.15.2
emoji==2.12.1
pyngrok==7.1.6
